Add out of range reporting
Many tablets send a specific sequence of bytes whenever a tool
leaves proximity. This change enables plugins and other internal
tools to be aware of these types of reports.

The included IntuosV1 change is working with a GD-0608-U, though it
may need testing with other tablets using the same parser.