Remove Configuration Editor
This has been a pain point for a while now and adds significant code
but almost nobody uses it and when they do it's just problems galore.

Not the fault of config editor, just that the only users of this feature
are those that don't know how to actually make a tablet configuration.
Those that do know just copies other configs as base and edits it as text.