Fix UC-Logic 1060N configuration.
No verification was ever done on pressure and considering the age of
this tablet I'm guessing it was plain wrong to begin with, so I've
corrected the value to 2047. See #1037 for the original implementation.

When the configuration was "fixed" originally, it ended up "fixing" it
to work with with a version with broken firmware. We should not be
accomodating configurations for broken firmware, as this leaves genuine
users in the dirt.